archive/backend/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
openai>=1.0.0
slowapi==0.1.9
structlog==24.1.0


